refactor(项目配置): 任务中心报告清理问题
