+++ backend/services/api-test/src/main/java/io/metersphere/api/service/CleanupTaskResultServiceImpl.java
@@ -51,6 +51,7 @@ public class CleanupTaskResultServiceImpl implements BaseCleanUpReport {
         }
         List<String> cleanTaskItemIds = execTaskItems.stream().map(ExecTaskItem::getId).toList();
         List<String> cleanIds = ListUtils.union(cleanTaskIds, cleanTaskItemIds);
+        LogUtils.info("清理当前项目[" + projectId + "]任务中心执行结果, 共[" + cleanIds.size() + "]条");
         if (CollectionUtils.isNotEmpty(cleanIds)) {
             // 清理任务-报告关系表
             SubListUtils.dealForSubList(cleanIds, 100, ids -> {
@@ -59,6 +60,6 @@ public class CleanupTaskResultServiceImpl implements BaseCleanUpReport {
                 apiReportRelateTaskMapper.deleteByExample(example);
             });
         }
-        LogUtils.info("清理当前项目[" + projectId + "]任务中心执行结果结束！");
+        LogUtils.info("清理当前项目[" + projectId + "]任务中心执行结果结束!");
     }
 }
